requirements.txt
GitPython == 3.1.43
PyQt6 == 6.7.0
